Queries funcs move to queries.py
