+++ scan/views/pools.py
@@ -1,5 +1,3 @@
-import json
-
 from django.db.models import F, OuterRef, Q
 from django.views.generic import ListView
 
@@ -11,7 +9,13 @@ from java_wallet.models import (
     Transaction,
 )
 from scan.caching_paginator import CachingPaginator
-from scan.helpers.queries import get_account_name
+from scan.helpers.queries import (
+    get_account_name,
+    get_count_forged_blocks_of_pool,
+    get_count_of_miners,
+    get_description_url,
+    get_timestamp_of_block,
+)
 from scan.views.base import IntSlugDetailView
 from scan.views.transactions import fill_data_transaction
 
@@ -21,57 +25,6 @@ def fill_data_pool(pool):
     pool["miners_cnt"] = get_count_of_miners(pool["pool_id"])
     pool["forged_block_cnt"] = get_count_forged_blocks_of_pool(pool["pool_id"])
     pool["block_timestamp"] = get_timestamp_of_block(pool["height"])
-
-
-def get_description_url(pool_id):
-    description = (
-        Account.objects.using("java_wallet")
-        .filter(id=pool_id)
-        .values_list("description", flat=True)
-        .filter(latest=1)
-        .first()
-    )
-    try:
-        return json.loads(description)["hp"]
-    except (json.JSONDecodeError, TypeError, KeyError):
-        return ''
-
-
-def get_count_of_miners(pool_id):
-    return (
-        RewardRecipAssign.objects.using("java_wallet")
-        .filter(recip_id=pool_id)
-        .filter(latest=1)
-    ).count()
-
-
-def get_timestamp_of_block(height):
-    return (
-        Block.objects.using("java_wallet")
-        .filter(height=height)
-        .values_list("timestamp", flat=True)
-        .first()
-    )
-
-
-def get_count_forged_blocks_of_pool(pool_id):
-    return (
-        RewardRecipAssign.objects.using("java_wallet")
-        .filter(~Q(recip_id=F('account_id')))
-        .annotate(
-            block=Block.objects.using("java_wallet")
-            .filter(generator_id=OuterRef("account_id"))
-            .order_by("-height")
-            .values("height")
-            [:1]
-        )
-        .order_by("-block")
-        .values("recip_id", "account_id", "block")
-        .exclude(block__isnull=True)
-        .exclude(recip_id__isnull=True)
-        .filter(latest=1)
-        .filter(recip_id=pool_id)
-    ).count()
 
 
 class PoolListView(ListView):
